+++ src/WhereGroup/MetadorBundle/Component/AddressInterface.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace WhereGroup\MetadorBundle\Component;
+
+use Symfony\Component\DependencyInjection\ContainerInterface;
+
+interface AddressInterface {
+    public function __construct(ContainerInterface $container);
+    public function get();
+    public function set($metadataObject);
+}