+++ src/WhereGroup/MetadorBundle/Component/MetadorUserInterface.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace WhereGroup\MetadorBundle\Component;
+
+use Symfony\Component\DependencyInjection\ContainerInterface;
+
+interface MetadorUserInterface {
+    public function __construct(ContainerInterface $container);
+}